# api/openapispec/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/variable-labels": {
            "get": {
                "description": "List variable labels",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variable_labels"
                ],
                "summary": "List variable labels",
                "operationId": "listVariableLabels",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Variable labels to filter variables by. Default to all(empty) labels.",
                        "name": "labels",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page number of the paginated list results. Default to 1.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page size of the paginated list results. If not set, the result won't be paginated.",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.VariableLabelsListResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create a new set of variable labels",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variable_labels"
                ],
                "summary": "Create variable labels",
                "operationId": "createVariableLabels",
                "parameters": [
                    {
                        "description": "Created variable labels",
                        "name": "variable_labels",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateVariableLabelsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/entity.VariableLabels"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
        "/api/v1/variables": {
            "get": {
                "description": "List variables",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variable"
                ],
                "summary": "List variables",
                "operationId": "listVariables",
                "parameters": [
                    {
                        "description": "Variable labels to filter variables by.",
                        "name": "variable",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ListVariableSetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/entity.VariableLabelsListResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create a new variable",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variable"
                ],
                "summary": "Create variable",
                "operationId": "createVariable",
                "parameters": [
                    {
                        "description": "Created variable",
                        "name": "variable",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateVariableSetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/entity.Variable"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v1/variables/{fqn}": {
            "get": {
                "description": "Get variable by variable fqn",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variable"
                ],
                "summary": "Get variable",
                "operationId": "getVariable",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Variable Fqn",
                        "name": "fqn",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/entity.Variable"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update the specified variable",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variable"
                ],
                "summary": "Update variable",
                "operationId": "updateVariable",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Variable Fqn",
                        "name": "fqn",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Variable",
                        "name": "variable",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateVariableRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/entity.Variable"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete a specified variable with fqn",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "variable"
                ],
                "summary": "Delete variable",
                "operationId": "deleteVariable",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Variable Fqn",
                        "name": "fqn",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.SecretStore": {
            "type": "object",
            "properties": {
                "provider": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "providerType": {
                    "type": "string"
                }
            }
        },
        "entity.SecretValue": {
            "type": "object",
            "properties": {
                "ref": {
                    "description": "Ref will only return with the update secret variable",
                    "type": "string"
                },
                "secretStore": {
                    "$ref": "#/definitions/entity.SecretStore"
                },
                "value": {
                    "description": "Value to store in the secret store.",
                    "type": "string"
                }
            }
        },
...
        "entity.Variable": {
            "type": "object",
            "properties": {
                "fqn": {
                    "description": "Fqn is the fully qualified name of the variable.",
                    "type": "string"
                },
                "labels": {
                    "description": "Labels clarifies the scope of the variable.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "type": {
                    "description": "Type is the type of the variable.",
                    "type": "string"
                },
                "value": {
                    "description": "Value is the value of the variable.",
                    "type": "string"
                },
                "variableKey": {
                    "description": "VariableKey is the access path for the variable.",
                    "type": "string"
                }
            }
        },
        "entity.VariableLabels": {
            "type": "object",
            "properties": {
                "labels": {
                    "description": "Labels is the list of variable labels, which should be sorted\nin ascending order of priority.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "variableKey": {
                    "description": "VariableKey is the access path for the variable.",
                    "type": "string"
                }
            }
        },
        "entity.VariableLabelsListResult": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "variableLabels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.VariableLabels"
                    }
                }
            }
        },
...
        "request.CreateVariableLabelsRequest": {
            "type": "object",
            "required": [
                "labels",
                "variableKey"
            ],
            "properties": {
                "labels": {
                    "description": "Labels is the list of variable labels, which should be sorted\nin ascending order of priority.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "variableKey": {
                    "description": "VariableKey is the access path for the variable.",
                    "type": "string"
                }
            }
        },
        "request.CreateVariableSetRequest": {
            "type": "object",
            "required": [
                "type",
                "value",
                "variableKey"
            ],
            "properties": {
                "labels": {
                    "description": "Labels clarifies the scope of the variable.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "type": {
                    "description": "Type is the type of the variable.",
                    "type": "string"
                },
                "value": {
                    "description": "Value is the value of the variable.",
                    "type": "string"
                },
                "variableKey": {
                    "description": "VariableKey is the access path for the variable.",
                    "type": "string"
                }
            }
        },
...
        "request.ListVariableSetRequest": {
            "type": "object",
            "required": [
                "labels"
            ],
            "properties": {
                "labels": {
                    "description": "Labels clarifies the scope of the variables.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "request.UpdateVariableRequest": {
            "type": "object",
            "properties": {
                "isSecret": {
                    "type": "boolean"
                },
                "key": {
                    "description": "key is the unique index to use value in specific stack",
                    "type": "string"
                },
                "path": {
                    "description": "Path is the relative path of the stack within the source.",
                    "type": "string"
                },
                "project": {
                    "description": "Project is the project related to stack",
                    "type": "string"
                },
                "secretValue": {
                    "$ref": "#/definitions/entity.SecretValue"
                },
                "value": {
                    "description": "value is the plain value of no sensitive data",
                    "type": "string"
                }
            }
        },
...
    }
}